raspberry-pi/requirements.txt
# Core dependencies
RPi.GPIO==0.7.1
adafruit-circuitpython-dht==3.7.0
adafruit-circuitpython-mcp3xxx==1.4.10
picamera==1.13
requests==2.28.2
python-dotenv==1.0.0
schedule==1.2.0
Pillow==9.5.0
numpy==1.24.3 